chore: forge fmt

+++ contracts/interfaces/IUniswapV2Pair.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.24;
 
 interface IUniswapV2Pair {
-    function approve(address spender, uint value) external returns (bool);
-    function balanceOf(address owner) external view returns (uint);
-    function transferFrom(address from, address to, uint value) external returns (bool);
+    function approve(address spender, uint256 value) external returns (bool);
+    function balanceOf(address owner) external view returns (uint256);
+    function transferFrom(address from, address to, uint256 value) external returns (bool);
 }
