+++ contracts/interfaces/IUniswapV2Router02.sol
@@ -4,16 +4,23 @@ pragma solidity ^0.8.24;
 interface IUniswapV2Router02 {
     function factory() external view returns (address);
     function addLiquidity(
-        address tokenA, address tokenB,
-        uint amountADesired, uint amountBDesired,
-        uint amountAMin, uint amountBMin,
-        address to, uint deadline
-    ) external returns (uint amountA, uint amountB, uint liquidity);
+        address tokenA,
+        address tokenB,
+        uint256 amountADesired,
+        uint256 amountBDesired,
+        uint256 amountAMin,
+        uint256 amountBMin,
+        address to,
+        uint256 deadline
+    ) external returns (uint256 amountA, uint256 amountB, uint256 liquidity);
 
     function removeLiquidity(
-        address tokenA, address tokenB,
-        uint liquidity,
-        uint amountAMin, uint amountBMin,
-        address to, uint deadline
-    ) external returns (uint amountA, uint amountB);
+        address tokenA,
+        address tokenB,
+        uint256 liquidity,
+        uint256 amountAMin,
+        uint256 amountBMin,
+        address to,
+        uint256 deadline
+    ) external returns (uint256 amountA, uint256 amountB);
 }
